Move classes to ArchivedTimelineV1 only
