+++ hudi-common/src/main/java/org/apache/hudi/common/table/timeline/versioning/v1/ArchivedTimelineV1.java
@@ -24,8 +24,10 @@ import org.apache.hudi.common.table.timeline.HoodieArchivedTimeline;
 import org.apache.hudi.common.table.timeline.HoodieInstant;
 import org.apache.hudi.common.table.timeline.HoodieInstantReader;
 import org.apache.hudi.common.table.timeline.HoodieTimeline;
+import org.apache.hudi.common.table.timeline.InstantComparison;
 import org.apache.hudi.common.util.Option;
 import org.apache.hudi.storage.StoragePath;
+import org.apache.hudi.storage.StoragePathInfo;
 
 import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.generic.IndexedRecord;
@@ -323,5 +325,43 @@ public class ArchivedTimelineV1 extends BaseTimelineV1 implements HoodieArchived
   @Override
   public HoodieArchivedTimeline reload(String startTs) {
     return new ArchivedTimelineV1(metaClient, startTs);
+  }
+
+  /**
+   * A time-based filter with range [startTs, endTs].
+   */
+  static class ClosedClosedTimeRangeFilter extends TimeRangeFilter {
+    private final String startTs;
+    private final String endTs;
+
+    public ClosedClosedTimeRangeFilter(String startTs, String endTs) {
+      super(startTs, endTs);
+      this.startTs = startTs;
+      this.endTs = endTs;
+    }
+
+    @Override
+    public boolean isInRange(String instantTime) {
+      return InstantComparison.isInClosedRange(instantTime, this.startTs, this.endTs);
+    }
+
+    public boolean isInRange(HoodieInstant instant) {
+      return InstantComparison.isInClosedRange(instant.requestedTime(), this.startTs, this.endTs);
+    }
+  }
+
+  /**
+   * A log file filter based on the full file paths
+   */
+  static class LogFileFilter {
+    private final Set<String> logFiles;
+
+    public LogFileFilter(Set<String> logFiles) {
+      this.logFiles = logFiles;
+    }
+
+    public boolean shouldLoadFile(StoragePathInfo storagePathInfo) {
+      return logFiles.contains(storagePathInfo.getPath().toString());
+    }
   }
 }
